Move awards to separate component.
