+++ src/components/common/Awards/index.module.css
@@ -1,0 +1,61 @@
+/* FONTS */
+@import url("https://fonts.googleapis.com/css2?family=Oswald:wght@200;300;400;500;600;700&family=Teko:wght@300;400;500;600;700&display=swap");
+/* FONT COLORS */
+/* BACKGROUND COLORS */
+/* HEIGHTS */
+/* WIDTHS */
+/* MARGINS */
+/** 12 division grid, 
+*   for multiples use calc($var*[const])
+*/
+/* ALWAYS USE max-width starting from xl down*/
+/* BREAKPOINTS */
+.root {
+  background-color: transparent;
+  margin: 0;
+  padding: 0;
+  border: none;
+}
+
+.container {
+  width: 800px;
+  margin: auto;
+  background-color: rgba(199, 226, 231, 0.9019607843);
+}
+
+.paraLeft,
+.paraRight {
+  padding: 20px;
+  position: relative;
+  width: 60%;
+  margin: 40px 0;
+  font-family: "Oswald", sans-serif;
+  line-height: 150%;
+  font-size: 25px;
+  z-index: 1;
+}
+.paraLeft::before,
+.paraRight::before {
+  z-index: -2;
+  position: absolute;
+  content: "";
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: blue;
+}
+
+.paraLeft {
+  margin-right: 40%;
+}
+.paraLeft::before {
+  background-color: #457b9d;
+}
+
+.paraRight {
+  margin-left: 40%;
+}
+.paraRight::before {
+  background-color: #1d3557;
+}/*# sourceMappingURL=index.module.css.map */
